+++ src/tests/outputs/sqlite.output.ts
@@ -1,4 +1,4 @@
-import { ColumnType } from "kysely";
+import type { ColumnType } from "kysely";
 
 export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
   ? ColumnType<S, I | undefined, U>